added basic camera functionality
okay this one's a long one.
We have two parts to this new class, CameraNetworkTable, which are the static parts and the instance parts.

The class parts allow us to initalize a new camera essentially by creating new "subtables" under the primary "CameraTable" NetworkTable. The primary CameraTable contains information such as:
1. The current VIEW camera (the camera that is being used by the monitor,)
2. The current IMAGING PROCESSING camera (that camera being used by the Jetson to calculate stuff.)
3. The sub tables for each camera

The current 6 static methods allow you to:
1. Cycle which camera is the "view" camera. The "view" camera shows up on the Driver's monitors.
2. Cycle which camera is the "processor" camera, which is the camera that is used for image processing.
3. Set the processor camera and view camera directly by passing in a camera itself (not used mostly, just use the instance's version)
4. find a camera by it's position
5. find a camera by it's http address.

The instance methods of CameraNetworkTable include getting the instance's network table, and calling direct "setAsProcessedCamera" and "setAsViewCamera" methods, which just call on the static methods and pass itself through the argument.

Soon to come:
1. More entries for settings
2. A verification method to make sure that the jetson is running image analysis on the camera you're trying to call on
3. Static methods to just get networktable entries directly from the chosen view camera or processor camera
4. methods to set specific entries and one that allows you to enter a string to index a certain entry and set it's value