CameraNetworkTables finished
<NOTE>: Requires extensive testing
CameraNetworkTables is essentially a talking platform to be used between the roboRIO and the Jetson, which controls camera processing.

First, by initializing a CameraNetworkTable,  you are setting up a NetworkTable with an ip address (the ip of the ip camera), and telling the Jetson to establish a connection with that ip camera.

Create a new table for Jetson and then create a subtable in that for the ip camera with the same ip name.

It will set up all of it's entries to the NetworkTable and then set a specific value within named "server_status_ready" to true. That is when you can setup the camera on your end, and then have a value in your Jetson folder that is client_status_ready, set that to true when you want to start.
There are two camera functions:
1. Processor camera
2. Vision camera
The vision camera will send its output to the driver laptop via gstreamer.
The processor camera will do anything else it wants to. A vision camera and a processor camera do not have to be the same camera, so if we have multiple cameras we can choose and assign the two roles to either one camera or split between both cameras.

The values you want to read are within Constants. For special variables like "xyzoom" which controls the continuous motion of the cameras, those specifically will only be for the VISION camera, and not the processor one. xyzoom will be under the MAIN TABLE, NOT A SPECIFIC ONE and will be ported directly to the vision camera. I can give you a screenshot of all the different modes for the string "options" that are present only with string-based settings. Number based settings are easy. The Average Joe Slider starts at 50 and goes between 0 and 100. Those are ported directly through a networktable entry and are just a single value. The booleans are special. Most of them are "requests," meaning you'll have to see they are turned on, complete the request, then turn them off.

I also added some widget classes for easier adding values to Shuffleboard. Most values on shuffleboard are editable. Values controlled by the controller are not changeable on Shuffleboard, but are still viewable.

Any questions refer to me please.
I also want to develop a back up using a USB-camera with a Microsoft Life cam or whatever.